Add the capability to use templates at compile time
Split the base app constructor into ::Index and ::Template classes and
render the template when the Rack middleware stack is constructed.

The template uses Mustache and can be populated by passing in a context object.

Aside from that, it works exactly the same as the regular static index.

This feature is useful for generating custom HTML content when the app starts up,
such as links and scripts with cache-busting hash strings.